apps/backend/pyproject.toml
[project]
name = "agentic-ecommerce-backend"
version = "0.1.0"
description = "Backend services for the agentic-ecommerce merchant agent."
requires-python = ">=3.13"
dependencies = [
  "beautifulsoup4==4.14.3",
  "faiss-cpu==1.13.2",
  "fastapi==0.128.0",
  "httpx==0.28.1",
  "langfuse>=2.0,<3.0",
  "lxml>=4.9.0",
  "neo4j==6.1.0",
  "numpy",
  "openai==2.16.0",
  "pandas>=2.0.0",
  "psycopg2-binary==2.9.11",
  "pydantic==2.12.5",
  "pytest==9.0.2",
  "pytest-asyncio>=0.23.0",
  "python-dotenv==1.2.1",
  "python-multipart",
  "redis==7.1.0",
  "requests==2.32.5",
  "sentence-transformers==5.2.2",
  "sqlalchemy==2.0.46",
  "streamlit>=1.36,<2.0",
  "uvicorn[standard]==0.40.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
where = ["."]
include = ["merchant_agent*", "agent*"]
